requirements.txt
fastapi~=0.95.1
pydantic[email]~=1.10.7
motor~=3.1.2
uvicorn~=0.22.0
passlib[bcrypt]~=1.7.4
python-jose[cryptography]~=3.3.0
python-multipart~=0.0.6